requirements.txt
streamlit==1.28.0
transformers==4.34.0
torch==2.0.1+cpu  # CPU-only PyTorch
sentencepiece==0.1.99
accelerate==0.24.0
sqlalchemy==2.0.20
python-dotenv==1.0.0
huggingface-hub==0.17.0
langchain==0.0.340
pandas==2.0.3